Remove `malachite-crypto` crate
