+++ code/Cargo.toml
@@ -5,7 +5,6 @@ members = [
   "crates/actors",
   "crates/cli",
   "crates/common",
-  "crates/crypto",
   "crates/consensus",
   "crates/driver",
   "crates/gossip-consensus",
@@ -43,7 +42,6 @@ malachite-actors           = { version = "0.1.0", path = "crates/actors" }
 malachite-cli              = { version = "0.1.0", path = "crates/cli" }
 malachite-common           = { version = "0.1.0", path = "crates/common" }
 malachite-consensus        = { version = "0.1.0", path = "crates/consensus" }
-malachite-crypto           = { version = "0.1.0", path = "crates/crypto" }
 malachite-driver           = { version = "0.1.0", path = "crates/driver" }
 malachite-gossip-consensus = { version = "0.1.0", path = "crates/gossip-consensus" }
 malachite-gossip-mempool   = { version = "0.1.0", path = "crates/gossip-mempool" }
