--- code/crates/crypto/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name                   = "malachite-crypto"
-description            = "Cyprogrography library for Malachite"
-version.workspace      = true
-edition.workspace      = true
-repository.workspace   = true
-license.workspace      = true
-publish.workspace      = true
-rust-version.workspace = true
-
-[features]
-ed25519 = ["dep:ed25519-consensus", "dep:signature", "dep:base64", "dep:rand", "dep:serde", "dep:sha3"]
-keccak256 = ["dep:sha3"]
-
-[dependencies]
-malachite-common.workspace = true
-
-base64    = { workspace = true, optional = true }
-rand      = { workspace = true, optional = true }
-serde     = { workspace = true, optional = true }
-sha3      = { workspace = true, optional = true }
-
-# Ed25519
-signature         = { workspace = true, optional = true }
-ed25519-consensus = { workspace = true, optional = true }
-
-[lints]
-workspace = true
-
-[package.metadata.docs.rs]
-all-features = true
-rustdoc-args = ["--cfg", "docsrs"]
