--- code/crates/crypto/src/lib.rs
@@ -1,7 +1,0 @@
-// For coverage on nightly
-#![allow(unexpected_cfgs)]
-#![cfg_attr(coverage_nightly, feature(coverage_attribute))]
-#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
-
-#[cfg(feature = "ed25519")]
-pub mod ed25519;
