+++ code/crates/test/Cargo.toml
@@ -17,10 +17,11 @@ malachite-node   = { workspace = true }
 malachite-round  = { workspace = true }
 malachite-vote   = { workspace = true }
 malachite-proto  = { workspace = true }
-malachite-crypto = { workspace = true, features = ["ed25519"] }
 
 async-trait        = { workspace = true }
+base64             = { workspace = true }
 bytesize           = { workspace = true }
+ed25519-consensus  = { workspace = true }
 hex                = { workspace = true }
 prost              = { workspace = true }
 prost-types        = { workspace = true }
